promises: rename Sending Transaction; fill in (#42)
* promises: rename Sending Transaction; fill in

Thoughts on naming this section:

- Max's original idea outline called it "Creating Transactions"
- Austin called it "Sending Transactions"
- sdk-rs reference docs call the same basic thing a "Promise"
  https://docs.rs/near-sdk/3.1.0/near_sdk/struct.Promise.html

I could also see calling it "Scheduling Transactions"

It seems super related to Cross Contract Calls, obviously.

I think a key concept we want to get across is that, whatever we call
these things, they are executed in the next block (or thereabouts), not
the same block as the function call that spawns them. So we want a name
that communicates this. (I guess we could also call it "Spawning
Transactions".)

Personally, I think maybe the best name for the sidebar section would be
"Promises" – nice and concise, and gets across the idea of "happens
later" for anyone already familiar with Promises from other languages.

To this end, I've renamed it "Promises" for now, though I'm happy to
change this if people feel strongly that a different name is better.
I've also ensured to use the words "creating" and "sending" in
combination with the word "transaction" throughout the copy of these
documents, so that if someone searches specifically for those keywords
they will find these pages.

I then added an intro explaining what Promises are, why they're
necessary, and what you can do with them.

Finally, the pages for Sending $Near, Creating Accounts, and Deploying
Contracts all have much more detail, including complete code examples
and explanations of all the parts I think might be tricky for NEAR
newbies.

Co-authored-by: Austin Abell <austinabell8@gmail.com>