Add react option to intro.md
I was going through the near sdk docs and first fired create-near-app and was super confused why NEAR decided to make the front end in HTML instead of react, which is the defacto way of building front end for most people. It wasn't until later that I realized that there was a flag to make it use react (or vue or angular). I think that it should 100% say that on the page since a large percentage of people exclusively use react now.