fixing the BIP-39 typo; also adding proof of where NEAR uses BIP-39
P.S. This chapter was cool. I didn't know this!

"There are no two words in [the BIP39 wordlist] with the same first 4 characters.

That means if you have the first 4 letters, you know the rest of the word by looking for those first 4 letters in the BIP39 wordlist. Some wallets will even fill in the rest of the word once the first 4 letters are entered."
...
This feature does not only apply to the English wordlist. It also applies to the Spanish, French, and Italian wordlists. 

https://www.blockplate.com/pages/first-4-letters-of-a-bip39-mnemonic-seed-phrase